Remove unnecessary TSDFKernel copy from cloud callback

+++ src/util/point.h
@@ -19,7 +19,7 @@ using Eigen::Matrix4i;
 using Eigen::Matrix4f;
 using Eigen::Quaternionf;
 using ScanPoints_t = std::vector<Vector3i>;
-using ScanPointType = int16_t;
+using ScanPointType = int32_t;
 using ScanPoint = Eigen::Matrix<ScanPointType, 3, 1>;
 
 /**
